+++ packages/rosen-extractor/lib/getRosenData/abstract/abstractRosenDataExtractor.ts
@@ -1,49 +1,45 @@
 import { TokenMap } from '@rosen-bridge/tokens';
 import { RosenData } from './types';
-import { AbstractLogger, DummyLogger } from '@rosen-bridge/abstract-logger';
+import { DummyLogger } from '@rosen-bridge/abstract-logger';
 import { validateAddress } from '@rosen-bridge/address-codec';
 
 export default abstract class AbstractRosenDataExtractor<TransactionType> {
   abstract readonly chain: string;
-  protected readonly logger: AbstractLogger;
-  protected readonly tokens: TokenMap;
-  protected readonly lockAddress: string;
 
-  constructor(lockAddress: string, tokens: TokenMap, logger?: AbstractLogger) {
-    this.lockAddress = lockAddress;
-    this.tokens = tokens;
-    this.logger = logger ? logger : new DummyLogger();
-  }
+  constructor(
+    protected readonly lockAddress: string,
+    protected readonly tokens: TokenMap,
+    protected readonly logger = new DummyLogger(),
+    protected readonly storeRawData = true,
+  ) {}
 
   /**
    * extracts RosenData from given lock transaction and wrap the amount
    */
   get = (transaction: TransactionType): RosenData | undefined => {
-    const rawData = this.extractRawData(transaction);
-    if (rawData) {
+    const data = this.extractData(transaction);
+    if (data) {
       try {
-        validateAddress(rawData.toChain, rawData.toAddress);
+        validateAddress(data.toChain, data.toAddress);
       } catch (e) {
         this.logger.debug(
-          `Receiver address validation failed (address [${rawData.toAddress}] on chain [${rawData.toChain}]) with error: ${e}`,
+          `Receiver address validation failed (address [${data.toAddress}] on chain [${data.toChain}]) with error: ${e}`,
         );
         return undefined;
       }
-      rawData.amount = this.tokens
-        .wrapAmount(
-          rawData.sourceChainTokenId,
-          BigInt(rawData.amount),
-          this.chain,
-        )
+      data.amount = this.tokens
+        .wrapAmount(data.sourceChainTokenId, BigInt(data.amount), this.chain)
         .amount.toString();
+      // Conditionally set rawData based on storeRawData flag
+      if (!this.storeRawData) {
+        data.rawData = 'raw-data extraction is off';
+      }
     }
-    return rawData;
+    return data;
   };
 
   /**
    * extracts RosenData from given lock transaction
    */
-  abstract extractRawData: (
-    transaction: TransactionType,
-  ) => RosenData | undefined;
+  abstract extractData: (transaction: TransactionType) => RosenData | undefined;
 }