+++ packages/rosen-extractor/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @rosen-bridge/rosen-extractor
 
+## 11.1.1
+
+### Patch Changes
+
+- Throw error when tx cbor output is not set on ogmios client
+
+## 11.1.0
+
+### Minor Changes
+
+- Add optional `storeRawData` parameter to control whether raw transaction data should exist in RosenData object
+
 ## 11.0.0
 
 ### Major Changes